+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 ### 7.4.10 (06/10/2020 - )
 * monitor: tweak kube monitor pod not ready error message
+* stats: track jvm_non_heap_used by default
 
 ### 7.4.9 (06/04/2020 - 06/09/2020)
 * http: for static content controllers, convert client abort exception to warning (still better use CDN/Storage solution to serve static resources)